+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2025-12-30
+
+### Changed
+
+- Add new step UpdateMavenWrapper to Java 25 migration with the following actions:
+    - update Maven Wrapper to latest version
+    - add a jvm.config file with options to ignore warning messages
+
 ## [0.2.4] - 2025-12-23
 
 ### Fixed